HAI-2309 removed 'user access rights' feature
