+++ .env
@@ -8,7 +8,6 @@ REACT_APP_SENTRY_DSN="https://24d073314fb3414cae1d6fef8c34a8f4@o394401.ingest.se
 REACT_APP_DISABLE_SENTRY=0
 REACT_APP_FEATURE_PUBLIC_HANKKEET=1
 REACT_APP_FEATURE_HANKE=1
-REACT_APP_FEATURE_ACCESS_RIGHTS=1
 REACT_APP_INFO_NOTIFICATION_LABEL=""
 REACT_APP_INFO_NOTIFICATION_LABEL_SV=""
 REACT_APP_INFO_NOTIFICATION_LABEL_EN=""
