+++ src/domain/hanke/hankeView/HankeView.tsx
@@ -210,14 +210,12 @@ const HankeView: React.FC<Props> = ({
         <Text tag="h2" styleAs="h3" weight="bold" spacingBottom="l">
           {hankeData?.hankeTunnus}
         </Text>
-        <FeatureFlags flags={['accessRights']}>
-          <Text tag="p" styleAs="body-s" spacingBottom="l">
-            <strong style={{ marginRight: 'var(--spacing-s)' }}>
-              {t('hankePortfolio:labels:oikeudet')}:
-            </strong>
-            {t(`hankeUsers:accessRightLevels:${signedInUser?.kayttooikeustaso}`)}
-          </Text>
-        </FeatureFlags>
+        <Text tag="p" styleAs="body-s" spacingBottom="l">
+          <strong style={{ marginRight: 'var(--spacing-s)' }}>
+            {t('hankePortfolio:labels:oikeudet')}:
+          </strong>
+          {t(`hankeUsers:accessRightLevels:${signedInUser?.kayttooikeustaso}`)}
+        </Text>
 
         <InformationViewHeaderButtons>
           <FeatureFlags flags={['hanke']}>
@@ -247,16 +245,14 @@ const HankeView: React.FC<Props> = ({
               ) : null}
             </CheckRightsByHanke>
           </FeatureFlags>
-          <FeatureFlags flags={['accessRights']}>
-            <Button
-              onClick={onEditRights}
-              variant="primary"
-              iconLeft={<IconUser aria-hidden="true" />}
-              theme="coat"
-            >
-              {t('hankeUsers:userManagementTitle')}
-            </Button>
-          </FeatureFlags>
+          <Button
+            onClick={onEditRights}
+            variant="primary"
+            iconLeft={<IconUser aria-hidden="true" />}
+            theme="coat"
+          >
+            {t('hankeUsers:userManagementTitle')}
+          </Button>
           <FeatureFlags flags={['hanke']}>
             <CheckRightsByHanke requiredRight="DELETE" hankeTunnus={hankeData.hankeTunnus}>
               <Button variant="primary" iconLeft={<IconCross aria-hidden="true" />} theme="black">
