+++ src/domain/hanke/portfolio/HankePortfolioComponent.tsx
@@ -228,7 +228,7 @@ const CustomAccordion: React.FC<CustomAccordionProps> = ({ hanke, signedInUser, 
                 {hanke.omistajat && hanke.omistajat[0]?.nimi}
               </Text>
             </div>
-            <FeatureFlags flags={['hanke', 'accessRights']}>
+            <FeatureFlags flags={['hanke']}>
               <div className={styles.gridBasicInfo}>
                 <Text tag="h3" styleAs="h6" weight="bold" className={styles.infoHeader}>
                   {t('hankeForm:labels:rights')}
@@ -295,12 +295,12 @@ const PaginatedPortfolio: React.FC<React.PropsWithChildren<PagedRowsProps>> = ({
     setHankeFilterEndDate,
   } = usePortfolioFilter();
 
-  const filterVaihe = (vaiheRows: Row[], id: string[], value: string[]) => {
+  const filterVaihe = (vaiheRows: Row[], _id: string[], value: string[]) => {
     if (value.length === 0) return vaiheRows;
     return vaiheRows.filter((hanke) => value.includes(hanke.values.vaihe));
   };
 
-  const filterTyyppi = (tyyppiRows: Row[], id: string[], value: string[]) => {
+  const filterTyyppi = (tyyppiRows: Row[], _id: string[], value: string[]) => {
     if (value.length === 0) return tyyppiRows;
     return tyyppiRows.filter((hanke) => {
       const includedTyypit = hanke.values.tyomaaTyyppi.filter((tyyppi: string) =>
@@ -311,7 +311,7 @@ const PaginatedPortfolio: React.FC<React.PropsWithChildren<PagedRowsProps>> = ({
   };
 
   const columns: Column[] = React.useMemo(() => {
-    const dateStartFilter = (dateStartRows: Row[], id: string[], dateStart: string) => {
+    const dateStartFilter = (dateStartRows: Row[], _id: string[], dateStart: string) => {
       if (dateStart) {
         if (hankeFilterEndDate) {
           return dateStartRows.filter((hanke) =>
@@ -326,7 +326,7 @@ const PaginatedPortfolio: React.FC<React.PropsWithChildren<PagedRowsProps>> = ({
       return dateStartRows;
     };
 
-    const dateEndFilter = (dateEndRows: Row[], id: string[], dateEnd: string) => {
+    const dateEndFilter = (dateEndRows: Row[], _id: string[], dateEnd: string) => {
       if (dateEnd) {
         if (hankeFilterStartDate) {
           return dateEndRows.filter((hanke) =>
