+++ src/domain/johtoselvitys/BasicInfo.tsx
@@ -1,101 +1,22 @@
-import React, { useEffect, useState } from 'react';
-import { Checkbox, Link, Notification, Select, SelectionGroup } from 'hds-react';
+import { Checkbox, Link, SelectionGroup } from 'hds-react';
 import { useFormContext } from 'react-hook-form';
 import { Trans, useTranslation } from 'react-i18next';
-import { isEqual } from 'lodash';
-import { ApplicationType, Contact, Customer, CustomerType } from '../application/types/application';
 import styles from './BasicInfo.module.scss';
 import TextInput from '../../common/components/textInput/TextInput';
 import TextArea from '../../common/components/textArea/TextArea';
 import Text from '../../common/components/text/Text';
-import ResponsiveGrid from '../../common/components/grid/ResponsiveGrid';
-import useUser from '../auth/useUser';
-import { findOrdererKey } from './utils';
 import { JohtoselvitysFormValues } from './types';
 import { getInputErrorText } from '../../common/utils/form';
 import BooleanRadioButton from '../../common/components/radiobutton/BooleanRadioButton';
-
-export interface InitialValueTypes {
-  applicationType: ApplicationType;
-  applicationData: {
-    applicationType: ApplicationType;
-    name: string;
-    startTime: number | null;
-    endTime: number | null;
-    identificationNumber: string; // hankeTunnus
-    clientApplicationKind: string;
-    workDescription: string;
-    constructionWork: boolean;
-    maintenanceWork: boolean;
-    emergencyWork: boolean;
-    propertyConnectivity: boolean;
-  };
-}
-
-export const initialValues: InitialValueTypes = {
-  applicationType: 'CABLE_REPORT',
-  applicationData: {
-    applicationType: 'CABLE_REPORT',
-    name: '',
-    startTime: null,
-    endTime: null,
-    identificationNumber: '',
-    clientApplicationKind: '',
-    workDescription: '',
-    constructionWork: false,
-    maintenanceWork: false,
-    emergencyWork: false,
-    propertyConnectivity: false,
-  },
-};
-
-const emptyCustomer: Customer = {
-  type: null,
-  name: '',
-  country: 'FI',
-  email: '',
-  phone: '',
-  registryKey: null,
-  ovt: null,
-  invoicingOperator: null,
-  sapCustomerNumber: null,
-};
-
-const emptyContact: Contact = {
-  email: '',
-  firstName: '',
-  lastName: '',
-  orderer: false,
-  phone: '',
-};
-
-const customerTypes: CustomerType[] = [
-  'customerWithContacts',
-  'contractorWithContacts',
-  'propertyDeveloperWithContacts',
-  'representativeWithContacts',
-];
-
-type Option = { value: CustomerType; label: string };
 
 export function BasicInfo() {
   const {
     register,
     watch,
-    setValue,
-    getValues,
     formState: { errors },
     trigger,
   } = useFormContext<JohtoselvitysFormValues>();
   const { t } = useTranslation();
-  const user = useUser();
-
-  const applicationSent: boolean = getValues('alluStatus') !== null;
-
-  const [selectedRole, setSelectedRole] = useState(() =>
-    // Set contact key with orderer field true to be the initial selected role.
-    findOrdererKey(getValues('applicationData')),
-  );
 
   const [
     constructionWorkChecked,
@@ -112,79 +33,6 @@ export function BasicInfo() {
   // Trigger validation for constructionWork field
   function validateConstructionWork() {
     trigger('applicationData.constructionWork');
-  }
-
-  useEffect(() => {
-    const userFirstName = user.data?.profile.given_name;
-    const userLastName = user.data?.profile.family_name;
-    if (userFirstName && !getValues(`applicationData.${selectedRole}.contacts.0.firstName`)) {
-      setValue(`applicationData.${selectedRole}.contacts.0.firstName`, userFirstName);
-    }
-    if (userLastName && !getValues(`applicationData.${selectedRole}.contacts.0.lastName`)) {
-      setValue(`applicationData.${selectedRole}.contacts.0.lastName`, userLastName);
-    }
-  }, [
-    user.data?.profile.given_name,
-    user.data?.profile.family_name,
-    setValue,
-    selectedRole,
-    getValues,
-  ]);
-
-  useEffect(() => {
-    if (
-      user.data?.profile.email &&
-      !getValues(`applicationData.${selectedRole}.contacts.0.email`)
-    ) {
-      setValue(`applicationData.${selectedRole}.contacts.0.email`, user.data?.profile.email);
-    }
-  }, [user.data?.profile.email, setValue, selectedRole, getValues]);
-
-  const roleOptions: Option[] = [
-    { value: 'customerWithContacts', label: t('form:yhteystiedot:titles:customerWithContacts') },
-    {
-      value: 'contractorWithContacts',
-      label: t('form:yhteystiedot:titles:contractorWithContacts'),
-    },
-    { value: 'propertyDeveloperWithContacts', label: t('form:yhteystiedot:titles:rakennuttajat') },
-    {
-      value: 'representativeWithContacts',
-      label: t('form:yhteystiedot:titles:representativeWithContacts'),
-    },
-  ];
-
-  function handleRoleChange(role: Option) {
-    if (role.value === selectedRole) return;
-
-    const previousRoleContacts = getValues(`applicationData.${selectedRole}.contacts`);
-    const contactToMove = previousRoleContacts.slice(0, 1);
-
-    let selectedRoleContacts = getValues(`applicationData.${role.value}.contacts`);
-
-    selectedRoleContacts =
-      selectedRoleContacts === null ||
-      (selectedRoleContacts.length === 1 && isEqual(selectedRoleContacts[0], emptyContact))
-        ? contactToMove
-        : contactToMove.concat(selectedRoleContacts);
-
-    // Remove moved contact from previous selected role contacts
-    setValue(
-      `applicationData.${selectedRole}.contacts`,
-      previousRoleContacts.length > 1 ? previousRoleContacts.slice(1) : [emptyContact],
-      {
-        shouldValidate: true,
-      },
-    );
-
-    if (!getValues(`applicationData.${role.value}.customer`)) {
-      setValue(`applicationData.${role.value}.customer`, emptyCustomer);
-    }
-
-    setValue(`applicationData.${role.value}.contacts`, selectedRoleContacts, {
-      shouldDirty: true,
-    });
-
-    setSelectedRole(role.value);
   }
 
   return (
@@ -219,6 +67,7 @@ export function BasicInfo() {
         label={t('hakemus:labels:nimi')}
         required
         autoComplete="on"
+        maxLength={100}
       />
 
       <TextInput
@@ -302,83 +151,8 @@ export function BasicInfo() {
         name="applicationData.workDescription"
         label={t('hakemus:labels:kuvaus')}
         required
+        maxLength={2000}
       />
-
-      <Text tag="h2" styleAs="h4" weight="bold" spacingBottom="s">
-        {t('form:labels:omatTiedot')}
-      </Text>
-
-      {applicationSent && (
-        <Notification
-          label={t('hakemus:notifications:sentApplicationPersonalInfoNotification')}
-          size="small"
-          style={{ marginBottom: 'var(--spacing-s)' }}
-        >
-          {t('hakemus:notifications:sentApplicationPersonalInfoNotification')}
-        </Notification>
-      )}
-
-      <ResponsiveGrid>
-        <Select<Option>
-          options={roleOptions}
-          id="roleInApplication"
-          defaultValue={roleOptions.find((role) => role.value === selectedRole)}
-          label={t('form:labels:role')}
-          onChange={handleRoleChange}
-          helper={t('form:labels:roleHelper')}
-          required
-          disabled={applicationSent}
-        />
-      </ResponsiveGrid>
-
-      {customerTypes.map((customerType) => {
-        if (customerType === selectedRole) {
-          return (
-            <React.Fragment key={customerType}>
-              <ResponsiveGrid>
-                <TextInput
-                  name={`applicationData.${selectedRole}.contacts.0.firstName`}
-                  label={t('hankeForm:labels:etunimi')}
-                  required
-                  readOnly={Boolean(user.data?.profile.given_name)}
-                  helperText={
-                    user.data?.profile.given_name ? t('form:labels:fromHelsinkiProfile') : ''
-                  }
-                  disabled={applicationSent}
-                />
-                <TextInput
-                  name={`applicationData.${selectedRole}.contacts.0.lastName`}
-                  label={t('hankeForm:labels:sukunimi')}
-                  required
-                  readOnly={Boolean(user.data?.profile.family_name)}
-                  helperText={
-                    user.data?.profile.family_name ? t('form:labels:fromHelsinkiProfile') : ''
-                  }
-                  disabled={applicationSent}
-                />
-              </ResponsiveGrid>
-
-              <ResponsiveGrid className={styles.formRow}>
-                <TextInput
-                  name={`applicationData.${customerType}.contacts.0.email`}
-                  label={t('form:yhteystiedot:labels:email')}
-                  required
-                  disabled={applicationSent}
-                  autoComplete="email"
-                />
-                <TextInput
-                  name={`applicationData.${customerType}.contacts.0.phone`}
-                  label={t('form:yhteystiedot:labels:puhelinnumero')}
-                  required
-                  disabled={applicationSent}
-                  autoComplete="tel"
-                />
-              </ResponsiveGrid>
-            </React.Fragment>
-          );
-        }
-        return null;
-      })}
     </div>
   );
 }
