--- src/domain/johtoselvitys/Contacts.module.scss
@@ -1,15 +1,0 @@
-.formInstructions {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-m);
-  margin-bottom: var(--spacing-s);
-}
-
-.fieldset {
-  padding: var(--spacing-s);
-  margin-bottom: var(--spacing-s);
-}
-
-.fillOwnInfoButton {
-  align-self: end;
-}
