--- src/domain/johtoselvitys/sampleRequest.json
@@ -1,116 +1,0 @@
-{
-  "id": null,
-  "userId": "JaskaJokunen",
-  "applicationType": "CABLE_REPORT",
-  "applicationData": {
-    "name": "Haitaton hankkeen nimi",
-    "customerWithContacts": {
-      "customer": {
-        "type": "COMPANY",
-        "name": "Haitaton Oy Ab",
-        "country": "FI",
-        "postalAddress": {
-          "streetAddress": {
-            "streetName": "Haittatie 6"
-          },
-          "postalCode": "12345",
-          "city": "Haitaton City"
-        },
-        "email": "info@haitaton.fi",
-        "phone": "042-555-6125",
-        "registryKey": "101010-FAKE",
-        "ovt": null,
-        "invoicingOperator": null,
-        "sapCustomerNumber": null
-      },
-      "contacts": [
-        {
-          "name": "Hannu Haitaton",
-          "postalAddress": {
-            "streetAddress": {
-              "streetName": "Haittatie 8"
-            },
-            "postalCode": "12345",
-            "city": "Haitaton City"
-          },
-          "email": "hannu@haitaton.fi",
-          "phone": "042-555-5216",
-          "orderer": true
-        }
-      ]
-    },
-    "geometry": {
-      "type": "GeometryCollection",
-      "crs": {
-        "type": "name",
-        "properties": {
-          "name": "EPSG:3879"
-        }
-      },
-      "geometries": [
-        {
-          "type": "Polygon",
-          "coordinates": [
-            [
-              [2.5495815e7, 6673160.0],
-              [2.5495855e7, 6673160.0],
-              [2.5495855e7, 6673190.0],
-              [2.5495815e7, 6673160.0]
-            ]
-          ]
-        }
-      ]
-    },
-    "startTime": 1646267516.878748,
-    "endTime": 1648081916.878748,
-    "pendingOnClient": false,
-    "identificationNumber": "HAI-123",
-    "clientApplicationKind": "Telekaapelin laittoa",
-    "workDescription": "Kaivuhommiahan nää tietty",
-    "contractorWithContacts": {
-      "customer": {
-        "type": "COMPANY",
-        "name": "Haitaton Oy Ab",
-        "country": "FI",
-        "postalAddress": {
-          "streetAddress": {
-            "streetName": "Haittatie 6"
-          },
-          "postalCode": "12345",
-          "city": "Haitaton City"
-        },
-        "email": "info@haitaton.fi",
-        "phone": "042-555-6125",
-        "registryKey": "101010-FAKE",
-        "ovt": null,
-        "invoicingOperator": null,
-        "sapCustomerNumber": null
-      },
-      "contacts": [
-        {
-          "name": "Kerttu Haitaton",
-          "postalAddress": {
-            "streetAddress": {
-              "streetName": "Haittatie 8"
-            },
-            "postalCode": "12345",
-            "city": "Haitaton City"
-          },
-          "email": "kerttu@haitaton.fi",
-          "phone": "042-555-2182",
-          "orderer": false
-        }
-      ]
-    },
-    "postalAddress": null,
-    "representativeWithContacts": null,
-    "invoicingCustomer": null,
-    "customerReference": null,
-    "area": null,
-    "propertyDeveloperWithContacts": null,
-    "constructionWork": true,
-    "maintenanceWork": false,
-    "emergencyWork": false,
-    "propertyConnectivity": false
-  }
-}
