+++ src/domain/johtoselvitys/validationSchema.ts
@@ -1,7 +1,11 @@
 import yup from '../../common/utils/yup';
-import isValidBusinessId from '../../common/utils/isValidBusinessId';
 import { JohtoselvitysFormValues } from './types';
-import { AlluStatus, ApplicationType, ContactType } from '../application/types/application';
+import { AlluStatus } from '../application/types/application';
+import {
+  applicationTypeSchema,
+  areaSchema,
+  customerWithContactsSchema,
+} from '../application/yupSchemas';
 
 const addressSchema = yup
   .object({
@@ -12,54 +16,6 @@ const addressSchema = yup
     city: yup.string(),
   })
   .nullable();
-
-const contactSchema = yup
-  .object({
-    firstName: yup.string().trim().required(),
-    lastName: yup.string().trim().required(),
-    email: yup.string().trim().email().max(100).required(),
-    phone: yup.string().trim().max(20).required(),
-    orderer: yup.boolean().required(),
-  })
-  .nullable()
-  .required();
-
-const customerSchema = contactSchema.omit(['firstName', 'lastName', 'orderer']).shape({
-  name: yup.string().trim().required(),
-  type: yup.mixed<ContactType>().nullable().required(),
-  registryKey: yup // business id i.e. Y-tunnus
-    .string()
-    .defined()
-    .nullable()
-    .when('type', {
-      is: (value: string) => value === 'COMPANY' || value === 'ASSOCIATION',
-      then: (schema) =>
-        schema.test('is-business-id', 'Is not valid business id', isValidBusinessId),
-      otherwise: (schema) => schema,
-    }),
-  country: yup.string().defined(),
-  ovt: yup.string().defined().nullable(),
-  invoicingOperator: yup.string().defined().nullable(),
-  sapCustomerNumber: yup.string().defined().nullable(),
-});
-
-const customerWithContactsSchema = yup.object({
-  customer: customerSchema.required(),
-  contacts: yup.array(contactSchema).defined(),
-});
-
-const areaSchema = yup.object({
-  geometry: yup.object({
-    type: yup.mixed<'Polygon'>().required(),
-    crs: yup.object({
-      type: yup.string().required(),
-      properties: yup.object({ name: yup.string() }),
-    }),
-    coordinates: yup.array().required(),
-  }),
-});
-
-const applicationTypeSchema = yup.mixed<ApplicationType>().defined().required();
 
 export const validationSchema: yup.ObjectSchema<JohtoselvitysFormValues> = yup.object({
   id: yup.number().defined().nullable(),
@@ -104,4 +60,12 @@ export const validationSchema: yup.ObjectSchema<JohtoselvitysFormValues> = yup.o
   }),
   selfIntersectingPolygon: yup.boolean().isFalse(),
   geometriesChanged: yup.boolean(),
+});
+
+export const newJohtoselvitysSchema = yup.object({
+  nimi: yup.string().trim().required(),
+  perustaja: yup.object({
+    sahkoposti: yup.string().email().required(),
+    puhelinnumero: yup.string().phone().max(20).required(),
+  }),
 });
