--- src/domain/johtoselvitys_new/BasicInfo.module.scss
@@ -1,10 +1,0 @@
-.formInstructions {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-s);
-  margin-bottom: var(--spacing-s);
-}
-
-.formRow {
-  margin-bottom: var(--spacing-l);
-}
