--- src/domain/johtoselvitys_new/Geometries.module.scss
@@ -1,34 +1,0 @@
-@import 'src/assets/styles/layout.scss';
-
-.mapContainer {
-  position: relative;
-  width: 100%;
-  height: 500px;
-  margin-bottom: var(--spacing-m);
-
-  &__inner {
-    position: absolute;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-  }
-}
-
-.areaInfoContainer {
-  padding: var(--spacing-m);
-  margin-top: var(--spacing-xs);
-  margin-bottom: var(--spacing-s);
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  gap: var(--spacing-s);
-
-  .nameInput {
-    width: auto;
-
-    @include respond-above(s) {
-      width: 350px;
-    }
-  }
-}
