--- src/domain/johtoselvitys_new/Geometries.tsx
@@ -1,337 +1,0 @@
-import { useEffect, useState } from 'react';
-import { Feature } from 'ol';
-import VectorSource, { VectorSourceEvent } from 'ol/source/Vector';
-import Polygon from 'ol/geom/Polygon';
-import Geometry from 'ol/geom/Geometry';
-import { useFieldArray, useFormContext } from 'react-hook-form';
-import { useTranslation } from 'react-i18next';
-import { Box, Flex } from '@chakra-ui/react';
-import { Button, Fieldset, IconAlertCircleFill, IconCross, Notification } from 'hds-react';
-import { debounce } from 'lodash';
-import { FeatureLike } from 'ol/Feature';
-
-import VectorLayer from '../../common/components/map/layers/VectorLayer';
-import Map from '../../common/components/map/Map';
-import FitSource from '../map/components/interations/FitSource';
-import Kantakartta from '../map/components/Layers/Kantakartta';
-import styles from './Geometries.module.scss';
-import Controls from '../../common/components/map/controls/Controls';
-import DrawModule from '../../common/components/map/modules/draw/DrawModule';
-import { formatSurfaceArea } from '../map/utils';
-import Text from '../../common/components/text/Text';
-import ResponsiveGrid from '../../common/components/grid/ResponsiveGrid';
-import DatePicker from '../../common/components/datePicker/DatePicker';
-import useLocale from '../../common/hooks/useLocale';
-import OverviewMapControl from '../../common/components/map/controls/OverviewMapControl';
-import useSelectableTabs from '../../common/hooks/useSelectableTabs';
-import useHighlightArea from '../map/hooks/useHighlightArea';
-import { JohtoselvitysArea, JohtoselvitysFormValues } from './types';
-import AddressSearchContainer from '../map/components/AddressSearch/AddressSearchContainer';
-import useAddressCoordinate from '../map/hooks/useAddressCoordinate';
-import { ApplicationGeometry } from '../application/types/application';
-import useForceUpdate from '../../common/hooks/useForceUpdate';
-import { getAreaDefaultName } from '../application/utils';
-import FeatureHoverBox from '../map/components/FeatureHoverBox/FeatureHoverBox';
-import ConfirmationDialog from '../../common/components/HDSConfirmationDialog/ConfirmationDialog';
-import LayerControl from '../../common/components/map/controls/LayerControl';
-import { MapTileLayerId } from '../map/types';
-import { useMapDataLayers } from '../map/hooks/useMapLayers';
-import Ortokartta from '../map/components/Layers/Ortokartta';
-import { HankeData } from '../types/hanke';
-import HankeLayer from '../map/components/Layers/HankeLayer';
-
-interface AreaToRemove {
-  index: number;
-  areaFeature: Feature<Geometry>;
-}
-
-function ErrorText({ children }: { children: string }) {
-  return (
-    <Box px="var(--spacing-l)" pb="var(--spacing-xl)" textAlign="center" color="var(--color-error)">
-      <Text tag="p">
-        <IconAlertCircleFill /> {children}
-      </Text>
-    </Box>
-  );
-}
-
-function getEmptyArea(feature: Feature<Geometry>): JohtoselvitysArea {
-  return {
-    geometry: new ApplicationGeometry((feature.getGeometry() as Polygon).getCoordinates()),
-    feature,
-  };
-}
-
-type Props = {
-  hankeData?: HankeData;
-};
-
-export function Geometries({ hankeData }: Readonly<Props>) {
-  const { t } = useTranslation();
-  const locale = useLocale();
-  const {
-    watch,
-    getValues,
-    setValue,
-    formState: { errors },
-  } = useFormContext<JohtoselvitysFormValues>();
-
-  const {
-    fields: applicationAreas,
-    append,
-    remove,
-  } = useFieldArray<JohtoselvitysFormValues, 'applicationData.areas'>({
-    name: 'applicationData.areas',
-  });
-
-  const [drawSource] = useState<VectorSource>(() => {
-    const features = applicationAreas.flatMap((area) => (area.feature ? area.feature : []));
-    return new VectorSource({ features });
-  });
-
-  const forceUpdate = useForceUpdate();
-
-  const addressCoordinate = useAddressCoordinate(
-    getValues('applicationData.postalAddress.streetAddress.streetName'),
-  );
-
-  const [featuresLoaded, setFeaturesLoaded] = useState(false);
-
-  const { mapTileLayers, toggleMapTileLayer } = useMapDataLayers();
-  const ortoLayerOpacity = mapTileLayers.kantakartta.visible ? 0.5 : 1;
-
-  useEffect(() => {
-    function handleAddFeature(e: VectorSourceEvent<FeatureLike>) {
-      if (e.feature) {
-        append(getEmptyArea(e.feature as Feature<Geometry>));
-      }
-    }
-
-    const handleChangeFeature = debounce(() => {
-      forceUpdate();
-      if (featuresLoaded) {
-        // When areas are modified set geometriesChanged with shouldDirty option
-        // so that when changing form page application is saved
-        setValue('geometriesChanged', true, { shouldDirty: true });
-      }
-    }, 100);
-
-    drawSource.on('addfeature', handleAddFeature);
-    drawSource.on('changefeature', handleChangeFeature);
-    drawSource.once('featuresloadstart', () => {
-      setFeaturesLoaded(true);
-    });
-
-    return function cleanup() {
-      handleChangeFeature.cancel();
-      drawSource.un('addfeature', handleAddFeature);
-      drawSource.un('changefeature', handleChangeFeature);
-    };
-  }, [drawSource, append, forceUpdate, setValue, featuresLoaded]);
-
-  const { tabRefs } = useSelectableTabs(applicationAreas, { selectLastTabOnChange: true });
-
-  const higlightArea = useHighlightArea();
-
-  const startTime = watch('applicationData.startTime');
-  const endTime = watch('applicationData.endTime');
-  const minEndDate = startTime ? new Date(startTime) : undefined;
-
-  const workTimesSet = startTime && endTime;
-
-  const [areaToRemove, setAreaToRemove] = useState<AreaToRemove | null>(null);
-
-  const [showSelfIntersectingNotification, setShowSelfIntersectingNotification] = useState(false);
-
-  function removeArea(index: number, areaFeature?: Feature<Geometry>) {
-    if (areaFeature !== undefined) {
-      setAreaToRemove({ index, areaFeature });
-    }
-  }
-
-  function confirmRemoveArea() {
-    if (areaToRemove !== null) {
-      const { index, areaFeature } = areaToRemove;
-      remove(index);
-      drawSource.removeFeature(areaFeature);
-      setAreaToRemove(null);
-    }
-  }
-
-  function closeAreaRemoveDialog() {
-    setAreaToRemove(null);
-  }
-
-  function handleSelfIntersectingPolygon(feature: Feature<Geometry> | null) {
-    if (feature) {
-      setValue('selfIntersectingPolygon', true);
-      setShowSelfIntersectingNotification(true);
-    } else {
-      setValue('selfIntersectingPolygon', false, { shouldValidate: true });
-      setShowSelfIntersectingNotification(false);
-    }
-  }
-
-  return (
-    <div>
-      <Text tag="p" spacingBottom="m">
-        {t('johtoselvitysForm:alueet:instructions1')}
-      </Text>
-      <Text tag="p" spacingBottom="s">
-        {t('johtoselvitysForm:alueet:instructions2')}
-      </Text>
-      <Text tag="p" spacingBottom="s">
-        {t('johtoselvitysForm:alueet:instructions3')}
-      </Text>
-      <Text tag="p" spacingBottom="m">
-        {t('form:requiredInstruction')}
-      </Text>
-
-      <Text tag="h3" styleAs="h4" weight="bold" spacingBottom="s">
-        {t('form:headers:alueet')}
-      </Text>
-
-      <Fieldset heading={t('form:labels:timespan')}>
-        <ResponsiveGrid>
-          <DatePicker
-            name="applicationData.startTime"
-            label={t('hakemus:labels:startDate')}
-            locale={locale}
-            required
-            helperText={t('form:helperTexts:dateInForm')}
-          />
-          <DatePicker
-            name="applicationData.endTime"
-            label={t('hakemus:labels:endDate')}
-            locale={locale}
-            required
-            minDate={minEndDate}
-            initialMonth={minEndDate}
-            helperText={t('form:helperTexts:dateInForm')}
-          />
-        </ResponsiveGrid>
-      </Fieldset>
-
-      <div className={styles.mapContainer}>
-        <Map zoom={9} center={addressCoordinate} mapClassName={styles.mapContainer__inner}>
-          {mapTileLayers.kantakartta.visible && <Kantakartta />}
-          {mapTileLayers.ortokartta.visible && <Ortokartta opacity={ortoLayerOpacity} />}
-
-          <AddressSearchContainer position={{ top: '1rem', left: '1rem' }} zIndex={101} />
-
-          {/* Don't show hanke areas when hanke is generated */}
-          {!hankeData?.generated && (
-            <HankeLayer
-              hankeData={hankeData && [hankeData]}
-              startDate={startTime?.toString() ?? hankeData?.alkuPvm}
-              endDate={endTime?.toString() ?? hankeData?.loppuPvm}
-            />
-          )}
-
-          <VectorLayer source={drawSource} zIndex={101} className="drawLayer" />
-
-          <FitSource source={drawSource} />
-
-          <OverviewMapControl />
-
-          <FeatureHoverBox
-            render={(featureWithPixel) => {
-              const areaName = featureWithPixel.feature.get('areaName');
-              return areaName ? <p>{areaName}</p> : null;
-            }}
-          />
-
-          <Controls>
-            {workTimesSet && (
-              <DrawModule
-                source={drawSource}
-                onSelfIntersectingPolygon={handleSelfIntersectingPolygon}
-              />
-            )}
-
-            <LayerControl
-              tileLayers={Object.values(mapTileLayers)}
-              onClickTileLayer={(id: MapTileLayerId) => toggleMapTileLayer(id)}
-            />
-          </Controls>
-        </Map>
-      </div>
-
-      {!workTimesSet && (
-        <Box px="var(--spacing-l)" py="var(--spacing-2-xl)" textAlign="center">
-          <Text tag="p">{t('johtoselvitysForm:alueet:giveDates')}</Text>
-        </Box>
-      )}
-
-      {errors.applicationData?.areas && <ErrorText>{t('form:errors:areaRequired')}</ErrorText>}
-
-      {showSelfIntersectingNotification && !errors.selfIntersectingPolygon && (
-        <Notification
-          type="error"
-          label={t('form:errors:selfIntersectingArea')}
-          notificationAriaLabel={t('common:components:notification:notification')}
-          autoClose={false}
-          dismissible
-          closeButtonLabelText={`${t('common:components:notification:closeButtonLabelText')}`}
-          onClose={() => setShowSelfIntersectingNotification(false)}
-          style={{ marginBottom: 'var(--spacing-s)' }}
-        />
-      )}
-
-      {errors.selfIntersectingPolygon && (
-        <ErrorText>{t('form:errors:selfIntersectingArea')}</ErrorText>
-      )}
-
-      <Text tag="h3" styleAs="h4" weight="bold">
-        {t('hakemus:labels:addedAreas')}
-      </Text>
-      <Box as="ul" paddingLeft="var(--spacing-l)">
-        {applicationAreas.map((area, index) => {
-          const geometry = area.feature?.getGeometry();
-          const surfaceArea = geometry && `(${formatSurfaceArea(geometry)})`;
-          const areaName = getAreaDefaultName(t, index, applicationAreas.length);
-
-          area.feature?.set('areaName', areaName);
-
-          return (
-            <li key={area.id}>
-              <Flex alignItems="center">
-                <Box
-                  as="button"
-                  type="button"
-                  _hover={{ textDecoration: 'underline' }}
-                  onClick={() => higlightArea(area.feature)}
-                >
-                  <div ref={tabRefs[index]}>
-                    {areaName} {surfaceArea}
-                  </div>
-                </Box>
-                <Button
-                  variant="supplementary"
-                  style={{ color: 'var(--color-error)' }}
-                  iconLeft={<IconCross aria-hidden="true" />}
-                  onClick={() => removeArea(index, area.feature)}
-                >
-                  {t('hankeForm:hankkeenAlueForm:removeAreaButton')}
-                </Button>
-              </Flex>
-            </li>
-          );
-        })}
-      </Box>
-
-      <ConfirmationDialog
-        title={t('hakemus:labels:removeAreaTitle')}
-        description={t('hakemus:labels:removeAreaDescription', {
-          areaName: getAreaDefaultName(t, areaToRemove?.index, applicationAreas.length),
-        })}
-        isOpen={areaToRemove !== null}
-        close={closeAreaRemoveDialog}
-        mainAction={confirmRemoveArea}
-        mainBtnLabel={t('common:confirmationDialog:confirmButton')}
-        variant="danger"
-      />
-    </div>
-  );
-}
-export default Geometries;
